UI: time display no longer bugs out after 2^31 half-samples. fix issue with Script700 viewer's heat map

+++ src/cli/play/SpcProgram/UI.cs
@@ -433,7 +433,7 @@ public static partial class CliMain {
 			}
 			
 			Display.ClearLine(Display.Height - 3);
-			Display.Write(formatTime((int) (curCycle / 32), TimeUnit.Timer2s), 0, Display.Height - 3, AnsiColor.Cyan);
+			Display.Write(formatTime(curCycle / 32, TimeUnit.Timer2s), 0, Display.Height - 3, AnsiColor.Cyan);
 			
 			var fullTimeInCycles = (long) (PrimaryEmu.SpcMetadata.LengthInSeconds ?? 60 * 12)  * 2048000;
 			var barLength        = Display.Width - 1 - 14;
